requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
python-dotenv==1.0.0
aiohttp==3.9.1
PyYAML==6.0.1